update junit engine cause maven was not running tests automatically

+++ src/test/java/services/indicators/MovingAverageIndicatorTest.java
@@ -23,14 +23,14 @@ class MovingAverageIndicatorTest {
 
         assertEquals(5, movingAverageTimeFrame.size());
         // first item is averaged with itself so the average is itself
-        assertEquals(BigDecimal.valueOf(2), movingAverageTimeFrame.getTicks().get(0).getValue());
+        assertEquals(BigDecimal.valueOf(2.0), BigDecimal.valueOf(movingAverageTimeFrame.getTicks().get(0).getValue().doubleValue()));
         // (4 + 2) / 2 = 3
-        assertEquals(BigDecimal.valueOf(3), movingAverageTimeFrame.getTicks().get(1).getValue());
+        assertEquals(BigDecimal.valueOf(3.0), BigDecimal.valueOf(movingAverageTimeFrame.getTicks().get(1).getValue().doubleValue()));
         // (8 + 4) / 2 = 6
-        assertEquals(BigDecimal.valueOf(6), movingAverageTimeFrame.getTicks().get(2).getValue());
+        assertEquals(BigDecimal.valueOf(6.0), BigDecimal.valueOf(movingAverageTimeFrame.getTicks().get(2).getValue().doubleValue()));
         // (9 + 8) / 2 = 8.5
-        assertEquals(BigDecimal.valueOf(8.5), movingAverageTimeFrame.getTicks().get(3).getValue());
+        assertEquals(BigDecimal.valueOf(8.5), BigDecimal.valueOf(movingAverageTimeFrame.getTicks().get(3).getValue().doubleValue()));
         // (1 + 9) / 2 = 5
-        assertEquals(BigDecimal.valueOf(5), movingAverageTimeFrame.getTicks().get(4).getValue());
+        assertEquals(BigDecimal.valueOf(5.0), BigDecimal.valueOf(movingAverageTimeFrame.getTicks().get(4).getValue().doubleValue()));
     }
 }
